+++ src/ScottPlot5/ScottPlot5/AxisRules/SquareZoomOut.cs
@@ -1,0 +1,30 @@
+﻿namespace ScottPlot.AxisRules;
+
+public class SquareZoomOut : IAxisRule
+{
+    readonly IXAxis XAxis;
+    readonly IYAxis YAxis;
+
+    public SquareZoomOut(IXAxis xAxis, IYAxis yAxis)
+    {
+        XAxis = xAxis;
+        YAxis = yAxis;
+    }
+
+    public void Apply(RenderPack rp)
+    {
+        double unitsPerPxX = XAxis.Width / rp.DataRect.Width;
+        double unitsPerPxY = YAxis.Height / rp.DataRect.Height;
+        double maxUnitsPerPx = Math.Max(unitsPerPxX, unitsPerPxY);
+
+        double halfHeight = rp.DataRect.Height / 2 * maxUnitsPerPx;
+        double yMin = YAxis.Range.Center - halfHeight;
+        double yMax = YAxis.Range.Center + halfHeight;
+        YAxis.Range.Set(yMin, yMax);
+
+        double halfWidth = rp.DataRect.Width / 2 * maxUnitsPerPx;
+        double xMin = XAxis.Range.Center - halfWidth;
+        double xMax = XAxis.Range.Center + halfWidth;
+        XAxis.Range.Set(xMin, xMax);
+    }
+}